fix notify_user
This PR fixes the `notify_user` job, as well as the
`check_for_changelog_` one.

In `check_for_changelog`, there was a bug where the comparison was made
between the _merge commit_ and the fork point, which means commits
lacking a changelog entry could appear successful if CI ran after master
had made progress.

So far, `notify_user` was only reporting on the status of
`collect_build_data`; however, over time, we have added a few other
jobs, and `collect_build_data` is not depending on all of them.
Therefore, this PR adds a direct dependency of `notify_user` on all the
other jobs, such that it is now always the last thing to run. This,
however, is not something we can guarantee moving forward, so we'll just
need to be careful when adding new jobs. :(

Finally, there is now way in Azure to query for the state of the
build, as opposed to the state of the current job. However, there is a
way to run a job only if all of its dependencies have succeeded, so we
introduce an intermediate `notify_user_helper` to allow us to report the
status of the entire build.

CHANGELOG_BEGIN
CHANGELOG_END